Add a function to retrieve exif offset within the file
While many RAW are direct TIFF containing the EXIF, this is not the case
for files like jpeg or fuji.

JPEG exif is located in the middle of the file and FUJI uses the embedded jpeg preview
to store the exif information.

Several exif metadata like the `Thumbnail Offset` refers to the offset compared to the root
of the exif data (i.e. number of bytes between the exif marker and the actual data).

Exposing those functions helps library users to compute the offset since the begining of the file

Change-Id: Id5f626055f80bb1db8cdfb49af368fd1fef79496